+++ packages/mobile/src/screens/home/components/product-section/styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  titleWrapper: {
+    flexWrap: 'wrap',
+  },
+});
+
+export { styles };
